Math Operations

+++ 02_data_types_and_variables/04a_math_operations/Source.cpp
@@ -11,7 +11,7 @@
 
 int main()
 {
-	int a{}, b{};
+	double a{}, b{};
 	std::cout << "Enter two values: ";
 	std::cin >> a >> b;
